Removed DiscordProvider context and hooks
