+++ models/User.ts
@@ -8,6 +8,14 @@ export interface Contributor extends User {
   role: Role;
 }
 
+export interface DiscordUser {
+  id: string
+  username: string
+  discriminator: string
+  avatar?: string
+  verified?: boolean
+}
+
 export interface LoggedInUser extends User {
   favorites: FavoritePage[];
   spaceRoles: SpaceRole []
