+++ prisma/schema.prisma
@@ -93,19 +93,20 @@ model Block {
 }
 
 model User {
-  id              String           @id @default(uuid()) @db.Uuid
-  createdAt       DateTime         @default(now())
-  updatedAt       DateTime         @default(now())
-  addresses       String[]
-  spaceRoles      SpaceRole[]
-  favorites       FavoritePage[]
-  pagePermissions PagePermission[]
-  pages           Page[]
-  blocks          Block[]
-  spaces          Space[]
-  inviteLinks     InviteLink[]
-  bounties        Bounty[]
-  applications    Application[]
+  id                    String           @id @default(uuid()) @db.Uuid
+  createdAt             DateTime         @default(now())
+  updatedAt             DateTime         @default(now())
+  addresses             String[]
+  spaceRoles            SpaceRole[]
+  favorites             FavoritePage[]
+  pagePermissions       PagePermission[]
+  pages                 Page[]
+  blocks                Block[]
+  spaces                Space[]
+  inviteLinks           InviteLink[]
+  bounties              Bounty[]
+  applications          Application[]
+  discord               Json?
 }
 
 model FavoritePage {
